requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.0.0
numpy==2.0.0
opencv-python-headless==4.10.0.84
tensorflow-cpu==2.18.0
anthropic==0.45.1
requests==2.32.3

# RAG Chatbot Dependencies
psycopg2~=2.9.10
psycopg2-binary~=2.9.11
torch~=2.6.0
python-dotenv~=1.2.1
langchain-google-genai~=3.2.0
langchain-core~=1.1.0
langchain-community~=0.4.1
langchain-postgres~=0.0.12
langchain-huggingface~=0.1.2
langsmith~=0.4.53
tqdm~=4.67.1
langchain-text-splitters~=1.0.0
sentence-transformers~=3.3.1
rank-bm25~=0.2.2
